fix: resolve TypeScript type errors with shared type definitions
- Create common task-types.ts with shared interfaces
- Remove duplicate type definitions from adapters
- Fix TaskHandler interface to use concrete types instead of generics
- Remove conflicting ProactiveGuidanceContext interfaces

Fixes compilation errors:
- TaskHandler type compatibility
- Import conflicts resolution
- Shared TaskRequest/TaskResponse types
